chore: use config in networking lib
