+++ networking/Cargo.toml
@@ -9,6 +9,7 @@ rusteron-client = { workspace = true }
 rusteron-media-driver = { workspace = true }
 tracing = { workspace = true }
 thiserror = { workspace = true }
+vex-config = { workspace = true }
 rand = "0.8"
 crossbeam = "0.8.4"
 dashmap = "6.1.0"
