+++ networking/tests/client_server_tests.rs
@@ -1,9 +1,8 @@
 use common::cmd::{decode_order_command, OrderCommand, OrderCommandType};
 use common::model::enums::{OrderAction, OrderType};
-use networking::client::config::GatewayConfig;
 use networking::client::{VexGateway, GatewayError};
-use networking::server::config::CoreConfig;
-use networking::server::server::VexCoreServer;
+use networking::server::VexCoreServer;
+use vex_config::{CoreNetworkingConfig, GatewayNetworkingConfig};
 use rusteron_client::{find_unused_udp_port, AeronFragmentHandlerCallback, AeronHeader};
 use tracing::{info, error};
 use std::time::Duration;
@@ -55,28 +54,13 @@ fn create_test_addresses() -> (SocketAddr, SocketAddr) {
 #[test_log::test]
 fn test_client_server_communication() {
     // This test demonstrates the actual usage of run() methods
-    let (server_addr, client_addr) = create_test_addresses();
+    let (server_addr, _client_addr) = create_test_addresses();
 
     // start aeron media driver in background, and create context directories for client and server in test-data directory
-    let mut current_dir = std::env::current_dir().unwrap();
-    current_dir.pop();
-    let context_path = current_dir.join("test-data").join("server");
-    let context_s = context_path.to_str().unwrap();
-    let context_path = current_dir.join("test-data").join("client");
-    let context_c = context_path.to_str().unwrap();
-
-    let context_c_clone = context_c.to_string();
     let client_handle = thread::spawn(move || -> Result<(), GatewayError> {
-        let client_config = GatewayConfig {
-            context_dir: context_c_clone,
-            local_address: client_addr.ip().to_string(),
-            core_address: server_addr.ip().to_string(),
-            core_port: server_addr.port(),
-            core_control_port: server_addr.port() + 1,
-            gateway_id: "gateway-1".to_string(),
-            max_message_size: 67,
-            enable_heartbeat: true,
-        };
+        let mut client_config = GatewayNetworkingConfig::test_defaults();
+        client_config.core_port = server_addr.port();
+        client_config.core_control_port = server_addr.port() + 1;
         info!("client_config: {:?}", client_config);
         let mut client = VexGateway::new(client_config)?;
 
@@ -111,23 +95,10 @@ fn test_client_server_communication() {
         Ok(())
     });
     
-    let context_s_clone = context_s.to_string();
     let _ = thread::spawn(move || {
-        let server_config = CoreConfig {
-            context_dir: context_s_clone,
-            local_address: server_addr.ip().to_string(),
-            initial_port: server_addr.port(),
-            initial_control_port: server_addr.port() + 1,
-            base_gateway_port: 40350,
-            max_gateways: 100,
-            max_connections_per_address: 10,
-            reserved_session_id_low: 0,
-            reserved_session_id_high: 2147483647,
-            enable_authentication: true,
-            enable_heartbeat: true,
-            gateway_timeout_seconds: 30,
-            core_id: "core-1".to_string(),
-        };
+        let mut server_config = CoreNetworkingConfig::test_defaults();
+        server_config.initial_port = server_addr.port();
+        server_config.initial_control_port = server_addr.port() + 1;
         info!("server_config: {:?}", server_config);
         let mut server = VexCoreServer::new(server_config).unwrap();                
         match server.start() {
